+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v2.3.2
+* Fixed some custom sprites reverting to default sprites after death
+* Removed `LoadSprites` config option
+
 ### v2.3.1
 * Fixed sprite loading breaking in certain areas
 * Fixed non-statically loaded audio not being replaced